requirements.txt
streamlit>=1.54.0
openai>=2.0.0
numpy>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
trafilatura>=2.0.0
